Split the query generation piece of BaseClusterIntegrationTest into another subclass so that it is possible to inherit from BaseClusterIntegrationTest without also inheriting the auto-generated query tests.
