+++ pinot-integration-tests/src/test/java/com/linkedin/pinot/integration/tests/OfflineClusterIntegrationTest.java
@@ -53,7 +53,7 @@ import static org.testng.Assert.assertEquals;
  * Integration test that converts Avro data for 12 segments and runs queries against it.
  */
 @Test
-public class OfflineClusterIntegrationTest extends BaseClusterIntegrationTest {
+public class OfflineClusterIntegrationTest extends BaseClusterIntegrationTestWithQueryGenerator {
   private static final Logger LOGGER = LoggerFactory.getLogger(OfflineClusterIntegrationTest.class);
 
   private final File _tmpDir = new File("/tmp/OfflineClusterIntegrationTest");
